device info added. Possibility to create another f log files was added. Operations with files repleced to FileManager.java file

+++ app/libs/PhoneLog/src/main/java/com/example/sdi/mylibrary/PhoneLog.java
@@ -1,6 +1,11 @@
 package com.example.sdi.mylibrary;
 
+import android.app.Activity;
+import android.content.Context;
+import android.graphics.Point;
+import android.os.Build;
 import android.os.Environment;
+import android.view.Display;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -15,7 +20,6 @@ import java.util.Calendar;
  */
 public class PhoneLog {
     public final static String WRITE_STATUS_OK = "OK";
-    private final static String mLogDirName = "Log";
 
     private static volatile PhoneLog mLog = null;
     private static String mLogFileName = "Log.log";
@@ -24,13 +28,35 @@ public class PhoneLog {
     private static FileOutputStream mLogFileOutputStream = null;
     private static OutputStreamWriter mOutputStreamWriter = null;
 
-    private static String mPrefix = "";
-
     private final static SimpleDateFormat mPrefixDateFormat = new SimpleDateFormat("hh:mm:ss(SSS) :\t");
 
+    private static FileManager mFileManager = null;
+
+    /**
+     * @param context: some activity from the application
+     * */
+    public void savePhoneSize( Activity context ){
+        append(DeviceInfoManager.getScreenSizeDescription(context));
+    }
+
+    /**
+     * 16
+     * */
+    public void saveSDKVersion(){
+        append(DeviceInfoManager.getSDKVersion());
+    }
+
+    /**
+     * example: Samsung galaxy star plus
+     * */
+    public void saveDevice() {
+        append(DeviceInfoManager.getDeviceModelName());
+    }
+
     private PhoneLog(){
-        // private constructorll;
-        createFile(isExternalStorageWritable(), true);
+        // private constructor;
+        mFileManager = new FileManager(); // .../Log by default
+        createFile(FileManager.isExternalStorageWritable(), true);
     }
 
     /**
@@ -58,6 +84,32 @@ public class PhoneLog {
     }
 
     /**
+     * delete all files from log folder
+     * */
+    public static void clearLogFolder(){
+        mFileManager.clearFolder();
+
+        // force creation of the new instance
+        mLog = null;
+    }
+
+    /**
+     * create new log file with name <<nameOfTheNewLog>> and delete previous if it needed
+     *
+     * @param nameOfTheNewLog: name.ext of new log file
+     * @param deletePreviousLogFile: previous log file will be deleted if TRUE
+    * */
+    public static void writeToNewNewLog(String nameOfTheNewLog, boolean deletePreviousLogFile){
+        if (mLogFile == null){
+            setLogName(nameOfTheNewLog);
+            createFile(FileManager.isExternalStorageWritable(), true);
+        } else {
+            setLogName(nameOfTheNewLog);
+            createFile(FileManager.isExternalStorageWritable(), deletePreviousLogFile);
+        }
+    }
+
+    /**
      * @param loggedMessage : message that will be added to log file
      * @return: status of write operation.
      *          must be OK.
@@ -65,7 +117,8 @@ public class PhoneLog {
     public String append(String loggedMessage){
         try {
             String formatedDate = mPrefixDateFormat.format(Calendar.getInstance().getTime());
-            mOutputStreamWriter.append(formatedDate + loggedMessage + "\n");
+            StringBuilder sb = new StringBuilder(formatedDate).append(loggedMessage).append("\n");
+            mOutputStreamWriter.append(sb.toString());
             mOutputStreamWriter.flush();
             return WRITE_STATUS_OK;
         } catch (IOException e) {
@@ -77,36 +130,28 @@ public class PhoneLog {
      * @param isExternalStorageAvailable: create file on sdCard if TRUE
      * @param deleteOldFile: file will be deleted if TRUE
      * */
-    private void createFile(boolean isExternalStorageAvailable, boolean deleteOldFile){
+    private static void createFile(boolean isExternalStorageAvailable, boolean deleteOldFile){
         if (isExternalStorageAvailable) {
             // write ot external storage
-            if (mLogFile == null) {
-                // create log file if not exists
-                File dir = new File(
-                        Environment.getExternalStorageDirectory().getAbsolutePath() +
-                        File.separator +
-                        mLogDirName
-                );
 
-                // create dir
-                dir.mkdirs();
-
-                // create file in dir
-                mLogFile = new File(dir, mLogFileName);
-                if (mLogFile.exists() && deleteOldFile){
-                    mLogFile.delete();
-                    try {
-                        mLogFile.createNewFile();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
+            // delete prev file if exists
+            if (mLogFile != null && mLogFile.exists() && deleteOldFile) {
+                mLogFile.delete();
             }
+
+            // create log file if not exists
+            mLogFile = mFileManager.createNewFile(isExternalStorageAvailable, mLogFileName);
+
+            try {
+                mLogFile.createNewFile();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
 
             try {
                 mLogFileOutputStream = new FileOutputStream(mLogFile);
                 mOutputStreamWriter = new OutputStreamWriter(mLogFileOutputStream);
-
             } catch (FileNotFoundException e) {
                 e.printStackTrace();
             }
@@ -114,11 +159,5 @@ public class PhoneLog {
             // sdCard not available
             // TODO: add possibility to write in internal storage
         }
-    }
-
-    /* Check if external storage is available for read and write*/
-    private boolean isExternalStorageWritable(){
-        String state = Environment.getExternalStorageState();
-        return Environment.MEDIA_MOUNTED.equals(state);
     }
 }
